poc: after several attempts to connect producer and consumer through actual mock servers and not through internal mocks I gave up for now -- too maintenance heavy, it's like building the suite that 1. Writes the pacts into JSON 2. Somehow brings the mock server from that JSON in build/pacts (one server per consumer test? our API arch can handle one per all files if it's possible) and has the URL stored in memory 3. Consumer test through the setup does the setup of that mock server if needed (tbd look how pact mock server even works)
I think that the best idea is to narrow-IT contract tests as we do with a DB -- we run consumer reaction to the right producer answers and we make sure that the producer reports the right answers.
Everything else could be derived from test review, if needed.
Better to have smaller cohesive tests on defined contracts than to try chaining a lot of interdependent interactions. Loose coupling >> trying to build an e2e based on pact contracts and testcontainers, HOWEVER TEMPTING IT MIGHT SEEM FROM THE FIRST GLANCE